- bugfixes - when button is displayed to player the entity now stays for a short amount of time to ease interaction (we can remove that later) - moved a lot of code from walkingentity to walkinganimal - this fixes a lot of troubles - even more stable now - still under development!
