+++ resources/config.yml
@@ -20,6 +20,9 @@ distances:
   interact: 6 # defines until which distance an interaction with an entity is possible
   find-partner: 49 # defines whe maximum distance when love mode is active and is searching for another partner
 
+interaction:
+  stay-time: 20 # ticks an entity waits for being sheared e.g.
+
 # control max spawning values ...
 max-spawn:
   bat: 2
