+++ src/revivalpmmp/pureentities/entity/monster/walking/Creeper.php
@@ -135,7 +135,9 @@ class Creeper extends WalkingMonster implements Explosive{
                 $this->motionX = $this->getSpeed() * 0.15 * ($x / $diff);
                 $this->motionZ = $this->getSpeed() * 0.15 * ($z / $diff);
             }
-            $this->yaw = rad2deg(-atan2($x / $diff, $z / $diff));
+            if ($diff > 0) {
+                $this->yaw = rad2deg(-atan2($x / $diff, $z / $diff));
+            }
             $this->pitch = $y == 0 ? 0 : rad2deg(-atan2($y, sqrt($x * $x + $z * $z)));
         }
